Implement the diagnostics
